requirements.txt
python-telegram-bot==20.3
openai
httpx==0.24.1